AllocateDRBDMinor per disk, not per instance
AllocateDRBDMinor used to allocate some DRBD minors for a given instance
using ComputeDRBDMap. ComputeDRBDMap didn't take into account, the DRBD
minors for disks that are detached (don't belong to any instance). This
patch fixes the above problem by modifying AllocateDRBDMinor and
ReleaseDRBDMinors in order to accept the disk for which we want to
allocate the minors (and not the instance). In addition, ComputeDRBDMap
now examines all the disks (even the detached ones) in order to compute
the minors.

Signed-off-by: Ilias Tsitsimpis <iliastsi@grnet.gr>