burnin: Test if attach/detach works properly
The attach/detach test is merged with the add/remove test. The test is
the following: an extra disk is added in each instance and is given a
random name. Once it's added successfully, we detach it and check if the
instance operates normally. Then, we pick instances randomly and attach
a detached disk to them, using an algorithm that matches a disk to an
instance based on its disk nodes. Then, we remove the disk.

Also, move the start/stop at the top of the tests, since it the
start/stop functionality is erroneously used first by other functions
(e.g. rename) and then tested on its own.