cmdlib: Return RPC result in AssembleInstanceDisks
The AssembleInstanceDisks function uses internally the
call_blockdev_assemble RPC call. It does two passes with this function,
the first in the secondary and the last in the primary node. The results
from the latter operation can be useful to the caller for hotplugging
reasons, so we return them in array format.