cmdlib: Add core attach/detach logic
Extend _ApplyContainerMods to accept an attach and detach function.
These functions operate in the same way as the add and remove functions,
minus the disk creation and disk removal parts.

Also, add two functions that should raise an exception if one attempts
to attach/detach a NIC.

Finally, update the tests in order to work with the new
_ApplyContainerMods function.